Split PhysSceneNode into more classes
